% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalThreadCount.R
\name{findOptimalThreadCount}
\alias{findOptimalThreadCount}
\title{Function that finds the optimal (fastest) thread count for different data.table functions}
\usage{
findOptimalThreadCount(rowCount, colCount)
}
\arguments{
\item{rowCount}{The number of rows in the data table.}

\item{colCount}{The number of columns in the data table.}
}
\value{
A data frame containing the optimal thread count for each data.table function.
}
\description{
This function finds the optimal thread count for running data.table functions with maximum efficiency.
}
\details{
Iteratively runs benchmarks with increasing thread counts and determines the optimal number of threads for each data.table function.
}
\examples{
\dontrun{
# Finding the best performing thread count for each benchmarked data.table function with a data size of 1000 rows and 10 columns:
optimalThreads <- findOptimalThreadCount(1000, 10)
}
}
