% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOptimalThreadCount.R
\name{setOptimalThreadCount}
\alias{setOptimalThreadCount}
\title{Function to set the optimal thread count for a specific \code{data.table} function}
\usage{
setOptimalThreadCount(benchmarkData, functionName)
}
\arguments{
\item{benchmarkData}{A \code{data.table} of class \code{data_table_threads_benchmark} containing benchmarked timings with corresponding thread counts.}

\item{functionName}{The name of the \code{data.table} function for which to set the optimal thread count.}
}
\value{
NULL.
}
\description{
Function to set the optimal thread count for a specific \code{data.table} function
}
\details{
Sets the thread count to the optimal value (fastest median runtime) for the specified \code{data.table} function based on the results obtained from \code{findOptimalThreadCount()}.
}
\examples{
\dontrun{
# Finding the best performing thread count for each benchmarked data.table function with a data size of 10000000 rows and 10 columns:
benchmarkData <- data.table.threads::findOptimalThreadCount(1e7, 10)
# Setting the optimal thread count for the 'forder' function:
setOptimalThreadCount(benchmarkData, "forder")
}
}
