% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setThreadCount.R
\name{setThreadCount}
\alias{setThreadCount}
\title{Function to set the thread count for a specific \code{data.table} function}
\usage{
setThreadCount(
  benchmarkData,
  functionName,
  type = "recommended",
  verbose = FALSE
)
}
\arguments{
\item{benchmarkData}{A \code{data.table} of class \code{data_table_threads_benchmark} containing benchmarked timings with corresponding thread counts.}

\item{verbose}{Option (logical) to enable or disable detailed message printing.}

\item{funcName}{The name of the \code{data.table} function for which to set the thread count.}
}
\value{
NULL.
}
\description{
Function to set the thread count for a specific \code{data.table} function
}
\details{
Sets the thread count to either the optimal (fastest median runtime) or recommended value (default) based on the chosen type argument for the specified \code{data.table} function based on the results obtained from \code{findOptimalThreadCount()}.
}
\examples{
\dontrun{
# Finding the best performing thread count for each benchmarked data.table function with a data size of 10000000 rows and 10 columns:
benchmarkData <- data.table.threads::findOptimalThreadCount(1e7, 10)
# Setting the optimal thread count for the 'forder' function:
setThreadCount(benchmarkData, "forder", "optimal")
getDTthreads()
# Setting the recommended thread count for the 'nafill' function:
setThreadCount(benchmarkData, "nafill", "recommended")
getDTthreads()
}
}
