% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarks.R
\name{runBenchmarks}
\alias{runBenchmarks}
\title{Function to run a set of predefined benchmarks for different data.table functions with varying thread counts}
\usage{
runBenchmarks(rowCount, colCount, threadCount)
}
\arguments{
\item{rowCount}{The number of rows in the data.table}

\item{colCount}{The number of columns in the data.table}

\item{threadCount}{The total number of threads to use.}
}
\value{
A data frame containing benchmarked timings for each data.table function with different thread counts.
}
\description{
Function to run a set of predefined benchmarks for different data.table functions with varying thread counts
}
\details{
Benchmarks various data.table functions that are parallelizable (setorder, GForce_sum, subsetting, frollmean, fcoalesce, between, fifelse, nafill, and CJ) with varying thread counts..
}
\examples{
\dontrun{
# Running a set of benchmarks for a data.table with 1000 rows and 10 columns, for thread counts going from 1 to the maximum number of available threads in the user's system:
benchmarkData <- runBenchmarks(1000, 10, getDTthreads())
}
}
